requirements.txt
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
drfdocs==0.0.11
olefile==0.44
Pillow==4.3.0
PyJWT==1.5.3
